Removed Office "unfurl" on main ballot page. Got "Back to" on candidate page to take voter back to office or ballot correctly.
