+++ src/js/components/Ballot/CandidateItemCompressed.jsx
@@ -1,9 +1,9 @@
 import React, { Component } from 'react';
+import { Link } from 'react-router';
 import PropTypes from 'prop-types';
 import TextTruncate from 'react-text-truncate';
 import { historyPush } from '../../utils/cordovaUtils';
 import ImageHandler from '../ImageHandler';
-import ItemSupportOpposeRaccoon from '../Widgets/ItemSupportOpposeRaccoon';
 import LearnMore from '../Widgets/LearnMore';
 import { renderLog } from '../../utils/logging';
 import OrganizationStore from '../../stores/OrganizationStore';
@@ -12,22 +12,16 @@ import VoterGuideStore from '../../stores/VoterGuideStore';
 
 export default class CandidateItemCompressed extends Component {
   static propTypes = {
-    link_to_ballot_item_page: PropTypes.bool,
     organization: PropTypes.object,
     oneCandidate: PropTypes.object,
-    showPositionStatementActionBar: PropTypes.bool,
   };
 
   constructor (props) {
     super(props);
     this.state = {
-      maximumNumberOfOrganizationsToDisplay: 4,
       organization: {},
     };
 
-    this.closeYourNetworkIsUndecidedPopover = this.closeYourNetworkIsUndecidedPopover.bind(this);
-    this.closeYourNetworkSupportsPopover = this.closeYourNetworkSupportsPopover.bind(this);
-    this.closeHighestIssueScorePopover = this.closeHighestIssueScorePopover.bind(this);
     this.getCandidateLink = this.getCandidateLink.bind(this);
     this.goToCandidateLink = this.goToCandidateLink.bind(this);
   }
@@ -93,25 +87,13 @@ export default class CandidateItemCompressed extends Component {
       return `/candidate/${oneCandidateWeVoteId}/btvg/${this.state.organization.organization_we_vote_id}`;
     } else {
       // If no organization_we_vote_id, signal that we want to link back to default ballot
-      return `/candidate/${oneCandidateWeVoteId}/b/btdb/`;
+      return `/candidate/${oneCandidateWeVoteId}/b/btdb/`; // back-to-default-ballot
     }
   }
 
   goToCandidateLink (oneCandidateWeVoteId) {
     const candidateLink = this.getCandidateLink(oneCandidateWeVoteId);
     historyPush(candidateLink);
-  }
-
-  closeYourNetworkSupportsPopover () {
-    this.refs['supports-overlay'].hide();
-  }
-
-  closeHighestIssueScorePopover () {
-    this.refs['highest-issue-score-overlay'].hide();
-  }
-
-  closeYourNetworkIsUndecidedPopover () {
-    this.refs['undecided-overlay'].hide();
   }
 
   render () {
@@ -121,10 +103,10 @@ export default class CandidateItemCompressed extends Component {
     }
 
     const oneCandidateWeVoteId = this.state.oneCandidate.we_vote_id;
-    const candidateSupportStore = SupportStore.get(oneCandidateWeVoteId);
-    const organizationsToFollowSupport = VoterGuideStore.getVoterGuidesToFollowForBallotItemIdSupports(oneCandidateWeVoteId);
-    const organizationsToFollowOppose = VoterGuideStore.getVoterGuidesToFollowForBallotItemIdOpposes(oneCandidateWeVoteId);
-    // console.log("OfficeItemCompressedRaccoon, just retrieved getVoterGuidesToFollowForBallotItemIdSupports");
+    // const candidateSupportStore = SupportStore.get(oneCandidateWeVoteId);
+    // const organizationsToFollowSupport = VoterGuideStore.getVoterGuidesToFollowForBallotItemIdSupports(oneCandidateWeVoteId);
+    // const organizationsToFollowOppose = VoterGuideStore.getVoterGuidesToFollowForBallotItemIdOpposes(oneCandidateWeVoteId);
+    // console.log("OfficeItemCompressed, just retrieved getVoterGuidesToFollowForBallotItemIdSupports");
     const candidatePartyText = this.state.oneCandidate.party && this.state.oneCandidate.party.length ? `${this.state.oneCandidate.party}. ` : '';
     const candidateDescriptionText = this.state.oneCandidate.twitter_description && this.state.oneCandidate.twitter_description.length ? this.state.oneCandidate.twitter_description : '';
     const candidateText = candidatePartyText + candidateDescriptionText;
@@ -133,7 +115,7 @@ export default class CandidateItemCompressed extends Component {
       <div key={oneCandidateWeVoteId} className="u-stack--md">
         <div className="o-media-object u-flex-auto u-min-50 u-push--sm u-stack--sm">
           {/* Candidate Photo, only shown in Desktop */}
-          <div onClick={this.props.link_to_ballot_item_page ? () => this.goToCandidateLink(this.state.oneCandidate.we_vote_id) : null}>
+          <Link to={this.getCandidateLink(this.state.oneCandidate.we_vote_id)}>
             <ImageHandler
               className="card-main__avatar-compressed o-media-object__anchor u-cursor--pointer u-self-start u-push--sm"
               sizeClassName="icon-candidate-small u-push--sm "
@@ -141,52 +123,25 @@ export default class CandidateItemCompressed extends Component {
               alt="candidate-photo"
               kind_of_ballot_item="CANDIDATE"
             />
-          </div>
+          </Link>
           <div className="o-media-object__body u-flex u-flex-column u-flex-auto u-justify-between">
             {/* Candidate Name */}
             <h4 className="card-main__candidate-name u-f5">
-              {this.props.link_to_ballot_item_page ? (
-                <a onClick={() => this.goToCandidateLink(this.state.oneCandidate.we_vote_id)}>
-                  <TextTruncate
-                    line={1}
-                    truncateText="…"
-                    text={this.state.oneCandidate.ballot_item_display_name}
-                    textTruncateChild={null}
-                  />
-                </a>
-              ) : (
+              <Link to={this.getCandidateLink(this.state.oneCandidate.we_vote_id)}>
                 <TextTruncate
                   line={1}
                   truncateText="…"
                   text={this.state.oneCandidate.ballot_item_display_name}
                   textTruncateChild={null}
                 />
-              )}
+              </Link>
             </h4>
             {/* Description under candidate name */}
             <LearnMore
               text_to_display={candidateText}
-              on_click={this.props.link_to_ballot_item_page ? () => this.goToCandidateLink(this.state.oneCandidate.we_vote_id) : null}
+              on_click={this.goToCandidateLink(this.state.oneCandidate.we_vote_id)}
               num_of_lines={3}
             />
-            {/* Organization's Followed AND to Follow Items */}
-            <div>
-              <div className="u-flex u-flex-auto u-flex-row u-justify-between u-items-center u-min-50">
-                {/* Positions in Your Network and Possible Voter Guides to Follow */}
-                <ItemSupportOpposeRaccoon
-                  ballotItemWeVoteId={oneCandidateWeVoteId}
-                  ballot_item_display_name={this.state.oneCandidate.ballot_item_display_name}
-                  display_raccoon_details_flag
-                  goToCandidate={() => this.goToCandidateLink(this.state.oneCandidate.we_vote_id)}
-                  maximumOrganizationDisplay={this.state.maximumNumberOfOrganizationsToDisplay}
-                  organizationsToFollowSupport={organizationsToFollowSupport}
-                  organizationsToFollowOppose={organizationsToFollowOppose}
-                  showPositionStatementActionBar={this.props.showPositionStatementActionBar}
-                  supportProps={candidateSupportStore}
-                  type="CANDIDATE"
-                />
-              </div>
-            </div>
           </div>
         </div>
       </div>
