+++ src/js/components/Navigation/HeaderBackToBar.jsx
@@ -190,7 +190,7 @@ class HeaderBackToBar extends Component {
     if (this.state.organizationWeVoteId && this.state.organizationWeVoteId !== '') {
       return `/office/${this.state.officeWeVoteId}/btvg/${this.state.organizationWeVoteId}`;
     } else {
-      return `/office/${this.state.officeWeVoteId}/b/btdb/`;
+      return `/office/${this.state.officeWeVoteId}/b/btdb/`; // back-to-default-ballot
     }
   }
 
@@ -256,7 +256,7 @@ class HeaderBackToBar extends Component {
       backToLink = '/ballot'; // Default to this
     }
 
-    if (this.props.params.back_to_variable === 'bto' || this.props.params.back_to_variable === 'btdo') {
+    if (this.props.params.back_to_variable === 'bto' || this.props.params.back_to_variable === 'btdo') { // back-to-default-office
       backToLink = this.getOfficeLink();
     }
 
@@ -267,7 +267,7 @@ class HeaderBackToBar extends Component {
       backToOrganizationLinkText = 'Back to Ballot';
     }
 
-    if (this.props.params.back_to_variable === 'bto' || this.props.params.back_to_variable === 'btdo') {
+    if (this.props.params.back_to_variable === 'bto' || this.props.params.back_to_variable === 'btdo') { // back-to-default-office
       if (this.state.officeName) {
         backToOrganizationLinkText = `Back to ${this.state.officeName}`;
       } else {
